+++ viburnum/cli/sqs_worker_template.py
@@ -1,0 +1,11 @@
+sqs_worker_template = """
+from viburnum.application import sqs_handler, SqsEventsSequence
+
+
+@sqs_handler("{sqs_name}")
+def {handler_name}(events: SqsEventsSequence):
+    for e in events:
+        print(e.message_id)
+        print(e.body)
+    e.delete()
+"""